+++ src/metrics/distance/overlap.js
@@ -10,6 +10,13 @@
  */
 
 /**
+ * Global sets used by the overlap function. This way, we don't need to
+ * create objects when computing the coefficient.
+ */
+let A = new Set(),
+    B = new Set();
+
+/**
  * Function returning the overlap coefficient between two sequences.
  *
  * @param  {mixed}  a     - The first sequence.
@@ -17,20 +24,36 @@
  * @return {number}       - The overlap coefficient between a & b.
  */
 export default function overlap(a, b) {
-
   if (a === b)
     return 1;
 
-  a = new Set(a);
-  b = new Set(b);
+  if (!a || !b)
+    return 0;
+
+  A.clear();
+  B.clear();
+
+  for (let i = 0, l = a.length; i < l; i++)
+    A.add(a[i]);
+  for (let i = 0, l = b.length; i < l; i++)
+    B.add(b[i]);
+
+  let tmp;
+
+  // Let's find the shortest set
+  if (A.size > B.size) {
+    tmp = A;
+    A = B;
+    B = tmp;
+  }
 
   // Computing intersection of both sets
-  const i = new Set();
+  let I = 0;
 
-  a.forEach(item => {
-    if (b.has(item))
-      i.add(item);
+  A.forEach(item => {
+    if (B.has(item))
+      I++;
   });
 
-  return i.size / Math.min(a.size, b.size);
+  return I / A.size;
 }